Move Category Abbreviations into livesplit-title-abbreviations
I'm not entirely sure if it makes sense to have them there, but it
doesn't really hurt.
